+++ app/jobs/update_news_job.rb
@@ -1,0 +1,7 @@
+class UpdateNewsJob < ApplicationJob
+  queue_as :news
+
+  def perform
+    News.active_state.where('expire_on < ?', Time.zone.today).update_all state: :expired, updated_at: Time.zone.now
+  end
+end